+++ QuoteCalculator/Models/LoanAllocation.cs
@@ -3,7 +3,7 @@
 	public class LoanAllocation
 	{
 		public string LenderName { get; set; }
-		public decimal Rate { get; set; }
-		public int Amount { get; set; }
+		public double Rate { get; set; }
+		public double Amount { get; set; }
 	}
 }
